Port majority of front end; see notes
Most of the front end has been copied over from the old implementaion
as of this commit. Note that the old app used an <iframe> element to show
page content in a theme, and that <iframe> had <style> tags with
internal rulesets.

For that reason, some minor typographical styles were excluded
(increase font size, weight, etc). Tweaking may be necessary later,
but no major damage was done.

Much of the CSS and HTML is tighly coupled and non-semantic (See issues #1 and #2).
Refactoring would be nice, but the front-end is working fine as it is. Leave it be
unless time can be set aside for polish.